Create FilesOlderthanXMins.ps1
This script is intended, given a directory path, to report the number of files that are > 5 minutes old.  The original use case was for a Windows SMTP Relay server where the queue directory was getting backed up because Office365 was rejecting emails.  The administrator stated that you could tell things were going wrong if you had old files in the queue directory.